feat(repay-scripts): manually add CoreJS polyfills and use automatic regenerator polyfills
CoreJS polyfills are needed to use REPAY's i18n components in IE.  However, babel's current
automatic polyfill functionality does not currently detect when it needs to be used for the Fluent
library, which is used by the i18n components.  Therefore we must add it manually.

BREAKING CHANGE: Polyfills for IE11 are now added to the bundles behind the scenes.

CACTUS-300