chore(repay-scripts): update dev dependencies
This update feature a breaking Jest changed that broke the resetModules behavior of Jest (I think
the issue is https://github.com/facebook/jest/issues/7863).  The only way to fix the associated test
was just to check that the module properly updated, and to eliminate the test of the original
content.  I think this is fine, as ultimately we really just care that the content updated properly.

CACTUS-298